requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.1.7
django-configurations==2.4
django-cors-headers==3.5.0
django-heroku==0.3.1
django-oauth-toolkit==2.2.0
djangorestframework==3.12.2
dnspython==2.3.0
gunicorn==20.0.4
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pymongo==4.3.3
python-dotenv==0.15.0
pytz==2022.7.1
requests==2.30.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.0.2
whitenoise==5.2.0
wrapt==1.15.0
